Refactor: move modules out of model subpackage; update imports
