--- bpod_rig/config/models/__init__.py
@@ -1,1 +1,0 @@
-"""Pydantic models for Bpod-rig settings"""
